Move `LinkOptions` to its own directory
