+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/LinkOptions/LinkOption.jsx
@@ -6,8 +6,8 @@ import Icon from "metabase/components/Icon";
 
 import { color } from "metabase/lib/colors";
 
-import { SidebarItemWrapper, SidebarItemStyle } from "./SidebarItem";
-import { SidebarIconWrapper } from "./ClickBehaviorSidebar.styled";
+import { SidebarItemWrapper, SidebarItemStyle } from "../SidebarItem";
+import { SidebarIconWrapper } from "../ClickBehaviorSidebar.styled";
 
 const LinkOption = ({ option, icon, onClick }) => (
   <SidebarItemWrapper onClick={onClick} style={{ ...SidebarItemStyle }}>
