+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/LinkOptions/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LinkOptions";
